Stop forcing evaluate
Why:

This was a safeguard against hyper-optimization using the test error, but
it is causing problem when we want to hyper-optimize using the training
error solely, without computing the validation error. The trials are
forced to validate by the space, and if this is disabled, the queries
does not find any trials because the experiment's query is forced with
validate=true.

Remove forced validate=true inside space and remove it inside
experiment's query.